Added a first draft for a Makefile.
At the moment we don't want to compile the library as a shared library. To
use this library, just copy the code in your project dir and write your
code for the data analysis and include the library as you would include
your own project code. Of cource you need to add your code files to the
makefile then.