Added the data class, which stores data of physical experiments.
The data class provides a simple interface to scientific data. The
data is read from a cassy file, when the object is created. It is stored
in a two dimensional double array and there are a few convenience
functions, which make data processing easier. More useful convenience
functions should be added in the future.