Added an early version of the LinearRegression class.
This class represents a linear regression. It computes the linear fit
and the residuals and pull distribution using root. It creates a
splitted canvas with a plot of the data with the fit, a plot of the
residuals and a plot of the pull distribution.

This class hasn't been tested yet and additionally functions will be
added soon. Hence, it's by no means complete.